+++ client/src/vscode-powerquery.api.d.ts
@@ -44,6 +44,7 @@ export interface LibraryFieldJson {
     readonly description: string | null;
 }
 
+// TODO: Rename and deprecate old API once the PQ SDK has been updated.
 export interface PowerQueryApi {
     readonly onModuleLibraryUpdated: (workspaceUriPath: string, library: LibraryJson) => void;
     readonly setLibrarySymbols: (librarySymbols: [string, LibraryJson | null][]) => Promise<void>;
